+++ backend/support/domain/src/main/java/kr/co/yigil/follow/infrastructure/FollowCountRepository.java
@@ -1,4 +1,4 @@
-package kr.co.yigil.follow.domain.repository;
+package kr.co.yigil.follow.infrastructure;
 
 import kr.co.yigil.follow.domain.FollowCount;
 import org.springframework.data.repository.CrudRepository;
